+++ queue/test/fiskaltrust.Middleware.Localization.QueueGR.UnitTest/Processors/DailyOperationsCommandProcessorGRTests.cs
@@ -1,32 +1,32 @@
 ﻿using fiskaltrust.Api.POS.Models.ifPOS.v2;
-using fiskaltrust.Middleware.Localization.QueueGR.GRSSCD;
 using fiskaltrust.Middleware.Localization.QueueGR.Processors;
 using fiskaltrust.Middleware.Localization.v2;
 using FluentAssertions;
-using Moq;
 using Xunit;
 using fiskaltrust.Middleware.Localization.v2.Models.ifPOS.v2.Cases;
 using Microsoft.Extensions.Logging;
+using Moq;
 
-namespace fiskaltrust.Middleware.Localization.QueueGR.UnitTest.QueueGR.Processors;
+namespace fiskaltrust.Middleware.Localization.QueueGR.UnitTest.Processors;
 
-public class ReceiptCommandProcessorGRTests
+public class DailyOperationsCommandProcessorGRTests
 {
+    private readonly ReceiptProcessor _sut = new(Mock.Of<ILogger<ReceiptProcessor>>(), null!, null!, new DailyOperationsCommandProcessorGR(), null!, null!);
+
     [Theory]
-    [InlineData(ReceiptCase.PointOfSaleReceipt0x0001)]
-    [InlineData(ReceiptCase.PaymentTransfer0x0002)]
-    [InlineData(ReceiptCase.PointOfSaleReceiptWithoutObligation0x0003)]
-    [InlineData(ReceiptCase.ECommerce0x0004)]
-    [InlineData(ReceiptCase.Protocol0x0005)]
+    [InlineData(ReceiptCase.ZeroReceipt0x2000)]
+    [InlineData(ReceiptCase.OneReceipt0x2001)]
+    [InlineData(ReceiptCase.ShiftClosing0x2010)]
+    [InlineData(ReceiptCase.DailyClosing0x2011)]
+    [InlineData(ReceiptCase.MonthlyClosing0x2012)]
+    [InlineData(ReceiptCase.YearlyClosing0x2013)]
     public async Task ProcessReceiptAsync_ShouldReturnEmptyList(ReceiptCase receiptCase)
     {
         var queue = TestHelpers.CreateQueue();
-        var queueGR = new Storage.GR.ftQueueGR();
-        var scuGR = new Storage.GR.ftSignaturCreationUnitGR();
         var queueItem = TestHelpers.CreateQueueItem();
         var receiptRequest = new ReceiptRequest
         {
-            ftReceiptCase = ((ReceiptCase) 0x4752_2000_0000_0000).WithCase(receiptCase)
+            ftReceiptCase = receiptCase
         };
         var receiptResponse = new ReceiptResponse
         {
@@ -38,27 +38,16 @@ public class ReceiptCommandProcessorGRTests
             ftReceiptIdentification = "receiptIdentification",
             ftReceiptMoment = DateTime.UtcNow,
         };
-        var grSSCDMock = new Mock<IGRSSCD>();
-        grSSCDMock.Setup(x => x.ProcessReceiptAsync(It.IsAny<ProcessRequest>()))
-            .ReturnsAsync(new ProcessResponse
-            {
-                ReceiptResponse = receiptResponse,
-            });
 
-        var receiptCommandProcessor = new ReceiptCommandProcessorGR(grSSCDMock.Object, queueGR, scuGR);
-        var receiptProcessor = new ReceiptProcessor(Mock.Of<ILogger<ReceiptProcessor>>(), null!, receiptCommandProcessor, null!, null!, null!);
-        var result = await receiptProcessor.ProcessAsync(receiptRequest, receiptResponse, queue, queueItem);
-
+        var result = await _sut.ProcessAsync(receiptRequest, receiptResponse, queue, queueItem);
         result.receiptResponse.Should().Be(receiptResponse);
         result.receiptResponse.ftState.Should().Be(0x4752_2000_0000_0000);
     }
 
     [Fact]
-    public async Task ProcessReceiptAsync_ShouldReturnError()
+    public async Task ProcessReceiptAsync_ShouldReturnError_IfInvalidCaseIsUsed()
     {
         var queue = TestHelpers.CreateQueue();
-        var queueGR = new Storage.GR.ftQueueGR();
-        var scuGR = new Storage.GR.ftSignaturCreationUnitGR();
         var queueItem = TestHelpers.CreateQueueItem();
         var receiptRequest = new ReceiptRequest
         {
@@ -74,11 +63,8 @@ public class ReceiptCommandProcessorGRTests
             ftReceiptIdentification = "receiptIdentification",
             ftReceiptMoment = DateTime.UtcNow,
         };
-        var grSSCDMock = new Mock<IGRSSCD>();
-        var receiptCommandProcessor = new ReceiptCommandProcessorGR(grSSCDMock.Object, queueGR, scuGR);
-        var receiptProcessor = new ReceiptProcessor(Mock.Of<ILogger<ReceiptProcessor>>(), null!, receiptCommandProcessor, null!, null!, null!);
-        var result = await receiptProcessor.ProcessAsync(receiptRequest, receiptResponse, queue, queueItem);
 
+        var result = await _sut.ProcessAsync(receiptRequest, receiptResponse, queue, queueItem);
         result.receiptResponse.Should().Be(receiptResponse);
         result.receiptResponse.ftState.Should().Be(0x4752_2000_EEEE_EEEE);
     }
